requirements.txt
Django==2.0.6
pytz==2018.4
reportlab
djangorestframework